Add global setting for sent media quality
Following in the shoes of Android and iOS, this change adds a global
setting for defining a preference for standard or high quality media
uploads in conversations.

This work takes inspiration from the theme and spell check settings that
are structurally similar (drop down select) and nearby in the ui (also
in Chats), respectively. Some patterns may be incorrect or dated, as
these were selected not based on how recently they were added.

No new features were added that needed novel testing pathways, we simply
wired up an existing boolean with a global default. As such,
functionality has been manually verified on a linux, and confirmed to
set the global default as desired. Some changes were also made to
Preferences.stories.tsx so that it would compile.