fix: make sure a convo exists before returning it in search results
There is an edge case if you are in the search results page, and delete
(or get a convo deleted from the network).
The corresponding convo does not exist anymore but the search result
selector still tries to extract values from it.

This commit fixes that issue by returning early if the corresponding
convo is not there anymore