correction of a wrong tense used in Czech sentence
I understand that this is not the ideal way to correct wrong translations in my native language, but trying to go through many forms and form posts is just a pain I don’t have the time for. I signed up for LQA via the form and wonder why there’s no translation management platform for such an influential open-source project in our everyday lives.