Fix active chat color when setting default color
When a conversation is selected and then the Preferences window
is used to update default Chat Color (Apperance -> Chat color), the
selected conversation messages do not get the updated chat color.

This was happening because a memoized selector cached the old color,
which is calculated in the getTimelineItem selector from the default
color setting.

This patch adds in SmartTimelineItem a wrapped getTimelineItem function
which takes a new argument for watched params for the memo cache. When
the watched params (in this case default color) change, the memo will
refresh the cached result of getTimelineItem correctly.

Fixes #6423