Add some tests demonstrating image-related behaviours.
[#6928] Transparent Png Doesn't Work (Even With Files)

Add a failing test to show that in some cases uploading a png file results in a jpg being
produced.

[#6881] File size increase when sending images

Add a passing test that automates the example written in the issue,
where a file's size increases after uploading.

Run with:

  npm run generate && npm run test-electron -- --grep="image processing"

It does add an extra devDependency, hope that's okay. It's for
inspecting jpeg images.