Use an absolute filename for the file save dialog on Linux
This fixes #6912.
On Gnome the save dialog works with any filename, on KDE Plasma
the save dialog opens with an empty filename input field,
unless the application uses an absolute file path instead of only the filename.
This commit will use the ~/Downloads directory as the default path,
which seems like a better choice then dumping files into the user's home dir.