Use an absolute filename for the file save dialog on Linux
This fixes #6912.
On Gnome the save dialog works with any filename, on KDE Plasma
the save dialog opens with an empty filename input field,
unless the application uses an absolute file path instead of only the filename.
The default save location is the Downloads directory in the user's home directory.